Updated Unity project
Added the "A" and "D" move functions into the move hero script.
Made a Prefab foulder for all our models (My wall is in there for now)
Created a parent class called "Humaniod".
And created a child class of Humaniod called "Hero"
Have tried to experiment with adding objects into the scene with code,
this have NOT been successful jet but the scrips are still there.