apple proximity and other stuff
Collision detector put on lars, values need to be assigned
Lars HP sat to 5, need changing when real HP value decided
Lars should now get HP when colliding with hearts and points
when colliding with seeds. gui for that needed.
Larses positions are locked to avoid him stepping out in the wrong
direction. Smoothening of the positioning needed.
apple proximity calculator made, although there is some problems with
the spawn of Larse due to the (Clone) naming. Needs to be fixed. If time
allows it, look for a simpler way to detect proximity.
Apples have been rezized, gotten a box collider and a rigid body. Their
positions are locked to avoid them stepping out of the coordinates in
which they are supposed to move though.