Steffen's homework done! (Read notes details!)
- GUI can not be seen.
This one is almost finished! But Mette is probably working with the
scenes so I will not finish the final details until nobody is working on
them!

- Tree shall be made solid
I found out that the three IS sloid, BUT its collider box is very small
and is INSIDE the three and not AROUND the tree (try to drag it out on
the scene and click on it, then will you see what I mean)

- You win / lose missing
They have been added and just need to be added into every scene when no
one is working on the scenes

-Tree root ranged fixed
The root's range have been doubled where its max range before was 5.0
and now is 10.0 (numbers might be fixed when the three gets into its
right position)

- Lars shall die when touching the tree
That have been done! But again shall he go INTO the three and and not
just touch the model itself.